我也不明白

+++ app/ucas/wheel/parallel.py
@@ -129,8 +129,8 @@ def kill_thread(thread=None, name=None, tid=0, exctype=SystemExit):
         tid = thread.ident
     if not tid:
         raise NotImplementedError('参数错误, 线程名, 线程id, 线程对象至少输入一个')
-    if not isinstance(tid, ctypes.c_long):
-        tid = ctypes.c_long(tid)
+    if not isinstance(tid, ctypes.c_longlong):
+        tid = ctypes.c_longlong(tid)
     if not inspect.isclass(exctype):
         raise TypeError("Only types can be raised (not instances)")
     res = ctypes.pythonapi.PyThreadState_SetAsyncExc(tid, ctypes.py_object(exctype))
@@ -150,7 +150,7 @@ def get_id(thread=None):
         pass
     else:
         raise NotImplementedError('参数错误, 这不是一个进程')
-    return ctypes.c_long(thread.ident)
+    return thread.ident
 
 
 def main():
